% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateData.R
\name{data.group}
\alias{data.group}
\title{Data generation for group lasso}
\usage{
data.group(N, p.group, beta.coef, sigma)
}
\arguments{
\item{N}{sample size}

\item{p.group}{number of covariates in each subgroup}

\item{beta.coef}{matrix of coefficient vectors (rows: groups, columns: coefficient values)}

\item{sigma}{model error variance}
}
\description{
Data generation for group lasso
}
