% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{group.group.lasso.computations}
\alias{group.group.lasso.computations}
\title{Title}
\usage{
group.group.lasso.computations(
  XX,
  response,
  index,
  index.subgroup,
  p.group,
  tau,
  delta.group = 2,
  delta.subgroup = 2,
  standardize = TRUE
)
}
\arguments{
\item{standardize}{}
}
\value{

}
\description{
Title
}
