% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{lasso.computations}
\alias{lasso.computations}
\title{Fit the Least Absolute Shrinkage and Selection Operator (LASSO)}
\usage{
lasso.computations(XX, response, delta, standardize)
}
\arguments{
\item{XX}{p by N matrix of predictors (N: sample size, p: number of predictors)}

\item{response}{1 by N matrix of response variable}

\item{delta}{Among the lasso solution path, the best descriptive model is the one which minimizes the loss function: (residual sum of squares)/(estimator of the model error variance) - (sample size) + delta*(number of predictors in the selected model). If delta = 2, this loss function is Mallows' Cp.}

\item{standardize}{logical. TRUE for standardizing the data.}
}
\value{
indicators of the selected predictors. 1 for selected predictors and 0 for not selected predictors
}
\description{
Fit the Least Absolute Shrinkage and Selection Operator (LASSO)
}
\examples{
set.seed(1)
XX <- matrix(rnorm(100*5, 0, 1),100,5)
response <- matrix(XX[,1] - XX[,2] + rnorm(100, 0, 1), 100)
predictors_selected <- lasso.computations(XX=XX,response=response,delta=2,standardize=TRUE)
}
