% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgslasso.R
\docType{package}
\name{sgslasso}
\alias{sgslasso}
\title{sgslasso package}
\description{
This package implements the sparse group-subgroup lasso which
identifies important regressor groups, subgroups, and individiual
variables.
}
\references{
Garcia, T.P., Müller, S., Carroll, R.J., and Walzem, R.L. (2013).
Identification of important regressor groups, subgroups, and individuals via regularization methods: application to gut
microbiome data. Bioinformatics, DOI: 10.1093/bioinformatics/btt608.
}
