% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{sparse.group.subgroup.computations}
\alias{sparse.group.subgroup.computations}
\title{Title}
\usage{
sparse.group.subgroup.computations(
  XX,
  response,
  group.index,
  subgroup.index,
  tau,
  alpha1,
  alpha2,
  alpha3,
  nlam,
  lambdas,
  lambda.accuracy,
  delta.group = 2,
  format.data = TRUE,
  cv.criterion = FALSE,
  nfold = 10,
  alphas.cv.range = seq(0.1, 0.95, by = 0.05)
)
}
\arguments{
\item{alphas.cv.range}{}
}
\value{

}
\description{
Title
}
